@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Company />
     <PackageId>Challonge-DotNet</PackageId>
     <Authors>Challonge-DotNet Contributors</Authors>
@@ -14,12 +14,12 @@
     <RepositoryType>GitHub</RepositoryType>
     <PackageProjectUrl>https://jacobhood.github.io/Challonge-DotNet/</PackageProjectUrl>
     <PackageReleaseNotes>Made Match.Optional nullable.</PackageReleaseNotes>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.8" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.8" />
   </ItemGroup>
 
 </Project>
