@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace Challonge.Objects
 {
-    internal class MatchAttachmentWrapper : ChallongeObjectWrapper<MatchAttachment>
-    {
-        [JsonProperty("match_attachment")]
-        internal override MatchAttachment Item { get; set; }
-    }
+	internal class MatchAttachmentWrapper : ChallongeObjectWrapper<MatchAttachment>
+	{
+		[JsonPropertyName("match_attachment")]
+		public override MatchAttachment Item { get; set; }
+	}
 }
