@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace Challonge.Objects
 {
-    internal class ParticipantWrapper : ChallongeObjectWrapper<Participant>
-    {
-        [JsonProperty("participant")]
-        internal override Participant Item { get; set; }
-    }
+	internal class ParticipantWrapper : ChallongeObjectWrapper<Participant>
+	{
+		[JsonPropertyName("participant")]
+		public override Participant Item { get; set; }
+	}
 }
