@@ -1,11 +1,10 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace Challonge.Objects
 {
-    internal class TournamentWrapper : ChallongeObjectWrapper<Tournament>
-    {
-        [JsonProperty("tournament")]
-        internal override Tournament Item { get; set; }
-
-    }
+	internal class TournamentWrapper : ChallongeObjectWrapper<Tournament>
+	{
+		[JsonPropertyName("tournament")]
+		public override Tournament Item { get; set; }
+	}
 }
