@@ -1,27 +1,6 @@
-﻿using Newtonsoft.Json;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Challonge.Objects
+﻿namespace Challonge.Objects
 {
-    public abstract class ChallongeObjectInfo
-    {
-        internal abstract Dictionary<string, object> ToDictionary(bool ignoreNulls);
-        private protected Dictionary<string, object> ToDictionaryWithKeyPrefix(string prefix, bool ignoreNulls)
-        {
-            JsonSerializerSettings settings = new()
-            {
-                DateTimeZoneHandling = DateTimeZoneHandling.Local
-            };
-
-            if (ignoreNulls)
-            {
-                settings.NullValueHandling = NullValueHandling.Ignore;
-            }
-
-            return JsonConvert.DeserializeObject<Dictionary<string, object>>(
-                JsonConvert.SerializeObject(this, settings))
-                .ToDictionary(kv => $"{prefix}[{kv.Key}]", kv => kv.Value);
-        }
-    }
+	public abstract class ChallongeObjectInfo
+	{
+	}
 }
