@@ -1,118 +1,117 @@
-﻿using Challonge.JsonConverters;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using System;
+﻿using System;
 using System.Collections.Generic;
+using System.Text.Json.Serialization;
+using Challonge.JsonConverters;
 
 namespace Challonge.Objects
 {
-    public class Match : ChallongeObject
-    {
-        [JsonProperty("id")]
-        public long Id { get; private set; }
+	public class Match : ChallongeObject
+	{
+		[JsonPropertyName("id")]
+		public long Id { get; init; }
 
-        [JsonProperty("tournament_id")]
-        public long TournamentId { get; private set; }
+		[JsonPropertyName("tournament_id")]
+		public long TournamentId { get; init; }
 
-        [JsonProperty("state")]
-        [JsonConverter(typeof(StringEnumConverter))]
-        public MatchState State { get; private set; }
+		[JsonPropertyName("state")]
+		public MatchState State { get; init; }
 
-        [JsonProperty("player1_id")]
-        public long? Player1Id { get; private set; }
+		[JsonPropertyName("player1_id")]
+		public long? Player1Id { get; init; }
 
-        [JsonProperty("player2_id")]
-        public long? Player2Id { get; private set; }
+		[JsonPropertyName("player2_id")]
+		public long? Player2Id { get; init; }
 
-        [JsonProperty("player1_prereq_match_id")]
-        public long? Player1PrereqMatchId { get; private set; }
+		[JsonPropertyName("player1_prereq_match_id")]
+		public long? Player1PrereqMatchId { get; init; }
 
-        [JsonProperty("player2_prereq_match_id")]
-        public long? Player2PrereqMatchId { get; private set; }
+		[JsonPropertyName("player2_prereq_match_id")]
+		public long? Player2PrereqMatchId { get; init; }
 
-        [JsonProperty("player1_is_prereq_match_loser")]
-        public bool Player1IsPrereqMatchLoser { get; private set; }
+		[JsonPropertyName("player1_is_prereq_match_loser")]
+		public bool Player1IsPrereqMatchLoser { get; init; }
 
-        [JsonProperty("player2_is_prereq_match_loser")]
-        public bool Player2IsPrereqMatchLoser { get; private set; }
+		[JsonPropertyName("player2_is_prereq_match_loser")]
+		public bool Player2IsPrereqMatchLoser { get; init; }
 
-        [JsonProperty("winner_id")]
-        public long? WinnerId { get; private set; }
+		[JsonPropertyName("winner_id")]
+		public long? WinnerId { get; init; }
 
-        [JsonProperty("loser_id")]
-        public long? LoserId { get; private set; }
+		[JsonPropertyName("loser_id")]
+		public long? LoserId { get; init; }
 
-        [JsonProperty("started_at")]
-        public DateTime? StartedAt { get; private set; }
+		[JsonPropertyName("started_at")]
+		public DateTime? StartedAt { get; init; }
 
-        [JsonProperty("created_at")]
-        public DateTime CreatedAt { get; private set; }
+		[JsonPropertyName("created_at")]
+		public DateTime CreatedAt { get; init; }
 
-        [JsonProperty("updated_at")]
-        public DateTime? UpdatedAt { get; private set; }
+		[JsonPropertyName("updated_at")]
+		public DateTime? UpdatedAt { get; init; }
 
-        [JsonProperty("identifier")]
-        public string Identifier { get; private set; }
+		[JsonPropertyName("identifier")]
+		public string Identifier { get; init; }
 
-        [JsonProperty("has_attachment")]
-        public bool HasAttachment { get; private set; }
+		[JsonPropertyName("has_attachment")]
+		public bool HasAttachment { get; init; }
 
-        [JsonProperty("round")]
-        public int Round { get; private set; }
+		[JsonPropertyName("round")]
+		public int Round { get; init; }
 
-        [JsonProperty("player1_votes")]
-        public int? Player1Votes { get; private set; }
+		[JsonPropertyName("player1_votes")]
+		public int? Player1Votes { get; init; }
 
-        [JsonProperty("player2_votes")]
-        public int? Player2Votes { get; private set; }
+		[JsonPropertyName("player2_votes")]
+		public int? Player2Votes { get; init; }
 
-        [JsonProperty("group_id")]
-        public long? GroupId { get; private set; }
+		[JsonPropertyName("group_id")]
+		public long? GroupId { get; init; }
 
-        [JsonProperty("attachment_count")]
-        public int? AttachmentCount { get; private set; }
+		[JsonPropertyName("attachment_count")]
+		public int? AttachmentCount { get; init; }
 
-        [JsonProperty("scheduled_time")]
-        public DateTime? ScheduledTime { get; private set; }
+		[JsonPropertyName("scheduled_time")]
+		public DateTime? ScheduledTime { get; init; }
 
-        [JsonProperty("location")]
-        public string Location { get; private set; }
+		[JsonPropertyName("location")]
+		public string Location { get; init; }
 
-        [JsonProperty("underway_at")]
-        public DateTime? UnderwayAt { get; private set; }
+		[JsonPropertyName("underway_at")]
+		public DateTime? UnderwayAt { get; init; }
 
-        [JsonProperty("optional")]
-        public bool? Optional { get; private set; }
+		[JsonPropertyName("optional")]
+		public bool? Optional { get; init; }
 
-        [JsonProperty("rushb_id")]
-        public long? RushbId { get; private set; }
+		[JsonPropertyName("rushb_id")]
+		public long? RushbId { get; init; }
 
-        [JsonProperty("completed_at")]
-        public DateTime? CompletedAt { get; private set; }
+		[JsonPropertyName("completed_at")]
+		public DateTime? CompletedAt { get; init; }
 
-        [JsonProperty("suggested_play_order")]
-        public int? SuggestedPlayOrder { get; private set; }
+		[JsonPropertyName("suggested_play_order")]
+		public int? SuggestedPlayOrder { get; init; }
 
-        [JsonProperty("forfeited")]
-        public bool? Forfeited { get; private set; }
+		[JsonPropertyName("forfeited")]
+		public bool? Forfeited { get; init; }
 
-        [JsonProperty("open_graph_image_file_name")]
-        public string OpenGraphImageFileName { get; private set; }
+		[JsonPropertyName("open_graph_image_file_name")]
+		public string OpenGraphImageFileName { get; init; }
 
-        [JsonProperty("open_graph_image_content_type")]
-        public string OpenGraphImageContentType { get; private set; }
+		[JsonPropertyName("open_graph_image_content_type")]
+		public string OpenGraphImageContentType { get; init; }
 
-        [JsonProperty("open_graph_image_file_size")]
-        public double? OpenGraphImageFileSize { get; private set; }
+		[JsonPropertyName("open_graph_image_file_size")]
+		public double? OpenGraphImageFileSize { get; init; }
 
-        [JsonProperty("prerequisite_match_ids_csv")]
-        [JsonConverter(typeof(PrerequisiteMatchIdsJsonConverter))]
-        public IEnumerable<long> PrerequisiteMatchIds { get; private set; }
+		[JsonPropertyName("prerequisite_match_ids_csv")]
+		[JsonConverter(typeof(PrerequisiteMatchIdsJsonConverter))]
+		public ICollection<long> PrerequisiteMatchIds { get; init; }
 
-        [JsonProperty("scores_csv")]
-        [JsonConverter(typeof(ScoresJsonConverter))]
-        public IEnumerable<Score> Scores { get; private set; }
+		[JsonPropertyName("scores_csv")]
+		[JsonConverter(typeof(ScoresJsonConverter))]
+		public ICollection<Score> Scores { get; init; }
 
-        internal Match() { }
-    }
+		[JsonConstructor]
+		internal Match() { }
+	}
 }
